Move ImageAnalysisOptions class to the `models` pacakge
