+++ sdk/vision/azure-ai-vision-imageanalysis/src/main/java/com/azure/ai/vision/imageanalysis/models/ImageAnalysisOptions.java
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.ai.vision.imageanalysis;
+package com.azure.ai.vision.imageanalysis.models;
 
 import java.util.List;
