+++ sdk/vision/azure-ai-vision-imageanalysis/src/samples/java/com/azure/ai/vision/imageanalysis/SampleAnalyzeAllImageFile.java
@@ -31,7 +31,6 @@
 
 import com.azure.ai.vision.imageanalysis.ImageAnalysisClient;
 import com.azure.ai.vision.imageanalysis.ImageAnalysisClientBuilder;
-import com.azure.ai.vision.imageanalysis.ImageAnalysisOptions;
 import com.azure.ai.vision.imageanalysis.models.CropRegion;
 import com.azure.ai.vision.imageanalysis.models.DenseCaption;
 import com.azure.ai.vision.imageanalysis.models.DetectedObject;
@@ -39,6 +38,7 @@ import com.azure.ai.vision.imageanalysis.models.DetectedPerson;
 import com.azure.ai.vision.imageanalysis.models.DetectedTag;
 import com.azure.ai.vision.imageanalysis.models.DetectedTextLine;
 import com.azure.ai.vision.imageanalysis.models.DetectedTextWord;
+import com.azure.ai.vision.imageanalysis.models.ImageAnalysisOptions;
 import com.azure.ai.vision.imageanalysis.models.ImageAnalysisResult;
 import com.azure.ai.vision.imageanalysis.models.VisualFeatures;
 import com.azure.core.credential.KeyCredential;
